+++ attributes/default.rb
@@ -8,6 +8,7 @@ default['kubernetes']['keep_versions']                      = 3
 default['kubernetes']['image']                              = 'gcr.io/google_containers/hyperkube'
 default['kubernetes']['interface']                          = 'eth1'
 default['kubernetes']['enable_firewall']                    = true
+default['kubernetes']['prometheus_operator_firewall']       = false
 default['kubernetes']['register_as']                        = 'ip'
 default['kubernetes']['proxy_mode']                         = 'iptables' # other possible values are: ipvs
 default['kubernetes']['use_sdn']                            = true